Update index.js & refactor
